logging-service/requirements.txt
Flask==2.1.0
hazelcast-python-client
